feat: use GoogleSafeBrowsing class
BREAKING CHANGE: Replaced functions with the GoogleSafeBrowsing class for cleaner and more structured code. Updated relevant tests and import statements to align with the new class-based implementation. Adjusted documentation in README to reflect these changes, providing examples and updated descriptions for using the new class.
