+++ .gitignore
@@ -6,3 +6,6 @@ node_modules/
 
 # macOS
 .DS_Store
+
+# environment variables
+.env*
