+++ README.md
@@ -1,5 +1,6 @@
 # google-safe-browsing
 
+[![JSR](https://jsr.io/badges/@hckhanh/google-safe-browsing)](https://jsr.io/@hckhanh/google-safe-browsing)
 [![Codacy Badge](https://app.codacy.com/project/badge/Grade/be4f5f8019a743f3878358399c110a36)](https://app.codacy.com/gh/hckhanh/google-safe-browsing/dashboard)
 
 A JavaScript client for [Google Safe Browsing](https://safebrowsing.google.com) [API](https://developers.google.com/safe-browsing)
@@ -9,7 +10,7 @@ A JavaScript client for [Google Safe Browsing](https://safebrowsing.google.com) 
 - [Zero dependencies](https://jsr.io/@hckhanh/google-safe-browsing/dependencies)
 - Built-in support for Edge runtime
 - Typesafe with TypeScript
-- Supports all Google Safe Browsing API v4 endpoints
+- Supports some common cases
 - Fully documented
 
 ## APIs
@@ -19,9 +20,10 @@ A JavaScript client for [Google Safe Browsing](https://safebrowsing.google.com) 
 Finds the threat entries that match the Safe Browsing lists.
 
 ```ts
-import { findThreadMatches } from '@hckhanh/google-safe-browsing'
+import { GoogleSafeBrowsing } from '@hckhanh/google-safe-browsing'
 
-const result = await findThreadMatches('apiKey', {
+const client = new GoogleSafeBrowsing('apiKey')
+const result = await client.findThreadMatches({
   client: {
     clientId: 'uniqueClientId',
     clientVersion: '1.0.0',
@@ -44,5 +46,9 @@ const hasRisk = result.matches !== undefined && result.matches.length > 0
 You can go to the [Releases](https://github.com/hckhanh/google-safe-browsing/releases) page to see the release notes.
 
 > [!NOTE]
-> The Safe Browsing API is for non-commercial use only. If you need to use APIs to detect malicious URLs for commercial
-> purposes - meaning 'for sale or revenue-generating purposes' - please refer to the [Web Risk API](https://github.com/hckhanh/google-web-risk).
+> Enables client applications to check web resources (most commonly URLs)
+> against Google-generated lists of unsafe web resources.
+> The Safe Browsing APIs are for non-commercial use only.
+> If you need to use APIs to detect malicious URLs for commercial purposes –
+> meaning “for sale or revenue-generating purposes” –
+> please refer to the [Web Risk API](https://github.com/hckhanh/google-web-risk).
