+++ package.json
@@ -7,12 +7,13 @@
   "sideEffects": false,
   "scripts": {
     "test": "vitest related src/*.ts",
-    "test:prettier": "prettier --check ."
+    "test:prettier": "prettier --check .",
+    "test:types": "deno check src/index.ts"
   },
   "author": {
     "name": "Khánh Hoàng",
-    "email": "hi@khanh.sh",
-    "url": "https://khanh.sh"
+    "email": "hi@khanh.id",
+    "url": "https://khanh.id"
   },
   "license": "MIT",
   "keywords": [
@@ -24,11 +25,11 @@
     "url"
   ],
   "devDependencies": {
-    "@edge-runtime/vm": "3.2.0",
-    "@types/node": "22.10.0",
-    "prettier": "3.2.5",
-    "typescript": "5.4.3",
-    "vitest": "1.3.1"
+    "@edge-runtime/vm": "4.0.4",
+    "@types/node": "22.10.1",
+    "prettier": "3.4.1",
+    "typescript": "5.7.2",
+    "vitest": "2.1.6"
   },
   "packageManager": "pnpm@9.14.3"
 }
