--- src/configs.ts
@@ -1,4 +1,0 @@
-/**
- * The base URL of the Google Safe Browsing API. The current version is v4.
- */
-export const endpoint = 'https://safebrowsing.googleapis.com/v4'
